backend/azure_functions/requirements.txt
azure-functions
requests
azure-storage-blob
pandas
python-dotenv
